fixed some bugs. moved apps.go to apps package.
